+++ Archive/Codeforces - Codeforces Round #641 (Div. 2)/C. Orac and LCM/solution.cpp
@@ -1,24 +1,24 @@
 #include <bits/stdc++.h>
- 
+
 #define pb push_back
 #define sz(v) ((int)(v).size())
 #define all(v) (v).begin(), (v).end()
 #define rf(i, a, b) for (int(i) = (a); (i) <= (b); (i)++)
 #define rb(i, b, a) for (int(i) = (b); (i) >= (a); (i)--)
- 
+
 using namespace std;
- 
+
 typedef long long ll;
 typedef vector<int> vi;
 typedef pair<int, int> ii;
 typedef pair<ii, int> iii;
 typedef vector<ii> vii;
- 
+
 const int N = 2e5 + 1;
 const int INF = 2e9;
- 
+
 int n, a[N], cnt[N];
-int ans = 1;
+ll ans = 1;
 int main() {
   ios_base::sync_with_stdio(0);
   cin.tie(0);
